+++ source/Knyaz.Optimus/Dom/Elements/HtmlCollection.cs
@@ -2,12 +2,12 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using Knyaz.Optimus.ScriptExecuting;
 
 namespace Knyaz.Optimus.Dom.Elements
 {
-	/// <summary>
-	/// Represents readonly collection of html elements.
-	/// </summary>
+	/// <summary> Represents readonly collection of html elements. </summary>
+	[JsName("HTMLCollection")]
 	public class HtmlCollection : IReadOnlyList<HtmlElement>
 	{
 		private readonly Func<IEnumerable<HtmlElement>> _items;
@@ -18,8 +18,17 @@ namespace Knyaz.Optimus.Dom.Elements
 
 		IEnumerator IEnumerable.GetEnumerator() { return GetEnumerator();}
 
+		[JsName("length")]
 		public int Count => _items().Count();
 
 		public HtmlElement this[int index] => _items().Skip(index).First();
+
+		/// <summary> Gets the item with specified Id. </summary>
+		public HtmlElement this[string id] => 
+			_items().FirstOrDefault(x => x.Id == id) 
+			?? (int.TryParse(id, out var idx) ? this[idx] : null);
+
+		/// <summary> Gets the item with specified Id. </summary>
+		public HtmlElement NamedItem(string id) => this[id];
 	}
 }